Set volume and size values as positive-only

+++ Alpaca.Markets/Messages/JsonAccount.cs
@@ -48,7 +48,7 @@ namespace Alpaca.Markets
         public Boolean ShortingEnabled { get; set; }
 
         [JsonProperty(PropertyName = "multiplier", Required = Required.Default)]
-        public Int64 Multiplier { get; set; }
+        public Byte Multiplier { get; set; }
 
         [JsonProperty(PropertyName = "buying_power", Required = Required.Always)]
         public Decimal BuyingPower { get; set; }
@@ -81,7 +81,7 @@ namespace Alpaca.Markets
         public Decimal LastMaintenanceMargin { get; set; }
 
         [JsonProperty(PropertyName = "daytrade_count", Required = Required.Default)]
-        public Int64 DayTradeCount { get; set; }
+        public UInt64 DayTradeCount { get; set; }
 
         [JsonProperty(PropertyName = "sma", Required = Required.Default)]
         public Decimal Sma { get; set; }
